8.0-reflextion-agent/pyproject.toml
[project]
name = "8-0-reflextion-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.3.0",
    "langchain-openai>=1.2.1",
    "langchain-tavily>=0.2.18",
    "langgraph>=1.2.0",
    "python-dotenv>=1.2.2",
]
